THE LAST 777 PATCH
FIXED ALL BUGS

+++ src/dataFuncs.h
@@ -14,7 +14,7 @@ void getFileSize( struct Fileinf* text );
 
 void fileToBuff( struct Fileinf* text );
 
-int parseBuff( struct Fileinf* text );
+size_t parseBuff( struct Fileinf* text );
 
 void freeFileinf( struct Fileinf* text );
 
@@ -22,6 +22,6 @@ void fillFileinf( struct Fileinf* text );
 
 void printFileinfBuff( struct Fileinf* text );
 
-int getNlines( struct Fileinf* text );
+size_t getNlines( struct Fileinf* text );
 
 void myBubbleSort( void* arr, size_t len, size_t elSize, int ( *cmp )( const void* a, const void* b ) );
